Refactor metrics files
